+++ trview.app/Settings/RandomizerSettings.cpp
@@ -142,7 +142,7 @@ namespace trview
 
     uint32_t RandomizerSettings::settings_of_type(Setting::Type type) const
     {
-        return std::count_if(settings.begin(), settings.end(),
-            [type](const auto& t) { return t.second.type == type; });
+        return static_cast<uint32_t>(std::count_if(settings.begin(), settings.end(),
+            [type](const auto& t) { return t.second.type == type; }));
     }
 }
